requirements.txt
Flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0